Lexical analysis
